refactor bit-install, move the logic from the command tsx file to a new class
