+++ src/extensions/install/install.provider.ts
@@ -2,6 +2,7 @@ import InstallCmd from './install.cmd';
 import { Workspace } from '../workspace';
 import { BitCli } from '../cli';
 import { PackageManager } from '../package-manager';
+import { Install } from './install';
 
 export type InstallConfig = {};
 
@@ -9,5 +10,6 @@ export type InstallDeps = [BitCli, Workspace, PackageManager];
 
 export async function provideInstaller(config: InstallConfig, [cli, workspace, packageManager]: InstallDeps) {
   // @ts-ignore
-  cli.register(new InstallCmd(workspace, packageManager));
+  const install = new Install(workspace, packageManager);
+  cli.register(new InstallCmd(install));
 }
